requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
clickhouse-connect==0.8.3
frozenlist==1.4.1
idna==3.10
logzero==1.7.0
lz4==4.3.3
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
pytz==2024.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
yarl==1.15.5
zstandard==0.23.0
